scheduler/requirements.txt
celery==3.1.23
redis
pymongo
python-dateutil

